+++ ideal/src/main/java/com/adyen/checkout/ideal/IdealComponent.kt
@@ -7,30 +7,83 @@
  */
 package com.adyen.checkout.ideal
 
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.adyen.checkout.action.core.internal.ActionHandlingComponent
 import com.adyen.checkout.action.core.internal.DefaultActionHandlingComponent
 import com.adyen.checkout.action.core.internal.ui.GenericActionDelegate
 import com.adyen.checkout.components.core.PaymentMethodTypes
+import com.adyen.checkout.components.core.internal.ButtonComponent
 import com.adyen.checkout.components.core.internal.ComponentEventHandler
 import com.adyen.checkout.components.core.internal.PaymentComponent
-import com.adyen.checkout.components.core.paymentmethod.IdealPaymentMethod
+import com.adyen.checkout.components.core.internal.PaymentComponentEvent
+import com.adyen.checkout.components.core.internal.toActionCallback
+import com.adyen.checkout.components.core.internal.ui.ComponentDelegate
+import com.adyen.checkout.core.AdyenLogLevel
+import com.adyen.checkout.core.internal.util.adyenLog
 import com.adyen.checkout.ideal.internal.provider.IdealComponentProvider
-import com.adyen.checkout.issuerlist.internal.IssuerListComponent
-import com.adyen.checkout.issuerlist.internal.ui.IssuerListDelegate
+import com.adyen.checkout.ideal.internal.ui.IdealDelegate
+import com.adyen.checkout.ui.core.internal.ui.ComponentViewType
+import com.adyen.checkout.ui.core.internal.ui.ViewableComponent
+import kotlinx.coroutines.flow.Flow
 
 /**
  * A [PaymentComponent] that supports the [PaymentMethodTypes.IDEAL] payment method.
  */
 class IdealComponent internal constructor(
-    delegate: IssuerListDelegate<IdealPaymentMethod, IdealComponentState>,
-    genericActionDelegate: GenericActionDelegate,
-    actionHandlingComponent: DefaultActionHandlingComponent,
-    componentEventHandler: ComponentEventHandler<IdealComponentState>,
-) : IssuerListComponent<IdealPaymentMethod, IdealComponentState>(
-    delegate,
-    genericActionDelegate,
-    actionHandlingComponent,
-    componentEventHandler,
-) {
+    private val idealDelegate: IdealDelegate,
+    private val genericActionDelegate: GenericActionDelegate,
+    private val actionHandlingComponent: DefaultActionHandlingComponent,
+    internal val componentEventHandler: ComponentEventHandler<IdealComponentState>,
+) : ViewModel(),
+    PaymentComponent,
+    ViewableComponent,
+    ButtonComponent,
+    ActionHandlingComponent by actionHandlingComponent {
+
+    @Suppress("ForbiddenComment")
+    // FIXME: Using actionHandlingComponent.activeDelegate will crash for QR code actions. This is a workaround for the
+    //  actual issue.
+    override val delegate: ComponentDelegate get() = genericActionDelegate.delegate
+
+    override val viewFlow: Flow<ComponentViewType?> = genericActionDelegate.viewFlow
+
+    init {
+        idealDelegate.initialize(viewModelScope)
+        genericActionDelegate.initialize(viewModelScope)
+        componentEventHandler.initialize(viewModelScope)
+    }
+
+    internal fun observe(
+        lifecycleOwner: LifecycleOwner,
+        callback: (PaymentComponentEvent<IdealComponentState>) -> Unit
+    ) {
+        idealDelegate.observe(lifecycleOwner, viewModelScope, callback)
+        genericActionDelegate.observe(lifecycleOwner, viewModelScope, callback.toActionCallback())
+    }
+
+    internal fun removeObserver() {
+        idealDelegate.removeObserver()
+        genericActionDelegate.removeObserver()
+    }
+
+    override fun setInteractionBlocked(isInteractionBlocked: Boolean) = Unit
+
+    @Suppress("FunctionOnlyReturningConstant")
+    @Deprecated("This method is only kept for backwards compatibility and will always return false.", ReplaceWith(""))
+    override fun isConfirmationRequired(): Boolean = false
+
+    @Deprecated("This method is only kept for backwards compatibility and doesn't do anything.", ReplaceWith(""))
+    override fun submit() = Unit
+
+    override fun onCleared() {
+        super.onCleared()
+        adyenLog(AdyenLogLevel.DEBUG) { "onCleared" }
+        idealDelegate.onCleared()
+        genericActionDelegate.onCleared()
+        componentEventHandler.onCleared()
+    }
 
     companion object {
         @JvmField