+++ ideal/src/test/java/com/adyen/checkout/ideal/IdealComponentTest.kt
@@ -1,0 +1,116 @@
+package com.adyen.checkout.ideal
+
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.viewModelScope
+import app.cash.turbine.test
+import com.adyen.checkout.action.core.internal.DefaultActionHandlingComponent
+import com.adyen.checkout.action.core.internal.ui.GenericActionDelegate
+import com.adyen.checkout.components.core.internal.ComponentEventHandler
+import com.adyen.checkout.components.core.internal.PaymentComponentEvent
+import com.adyen.checkout.ideal.internal.ui.IdealDelegate
+import com.adyen.checkout.test.LoggingExtension
+import com.adyen.checkout.test.TestDispatcherExtension
+import com.adyen.checkout.test.extensions.invokeOnCleared
+import com.adyen.checkout.ui.core.internal.test.TestComponentViewType
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.test.runTest
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertNull
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
+import org.mockito.Mock
+import org.mockito.junit.jupiter.MockitoExtension
+import org.mockito.kotlin.any
+import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.eq
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.whenever
+
+@ExtendWith(MockitoExtension::class, TestDispatcherExtension::class, LoggingExtension::class)
+internal class IdealComponentTest(
+    @Mock private val idealDelegate: IdealDelegate,
+    @Mock private val genericActionDelegate: GenericActionDelegate,
+    @Mock private val actionHandlingComponent: DefaultActionHandlingComponent,
+    @Mock private val componentEventHandler: ComponentEventHandler<IdealComponentState>,
+) {
+
+    private lateinit var component: IdealComponent
+
+    @BeforeEach
+    fun before() {
+        whenever(genericActionDelegate.viewFlow) doReturn MutableStateFlow(null)
+
+        component = IdealComponent(
+            idealDelegate,
+            genericActionDelegate,
+            actionHandlingComponent,
+            componentEventHandler,
+        )
+    }
+
+    @Test
+    fun `when component is created then delegates are initialized`() {
+        verify(idealDelegate).initialize(component.viewModelScope)
+        verify(genericActionDelegate).initialize(component.viewModelScope)
+        verify(componentEventHandler).initialize(component.viewModelScope)
+    }
+
+    @Test
+    fun `when component is cleared then delegates are cleared`() {
+        component.invokeOnCleared()
+
+        verify(idealDelegate).onCleared()
+        verify(genericActionDelegate).onCleared()
+        verify(componentEventHandler).onCleared()
+    }
+
+    @Test
+    fun `when observe is called then observe in delegates is called`() {
+        val lifecycleOwner = mock<LifecycleOwner>()
+        val callback: (PaymentComponentEvent<IdealComponentState>) -> Unit = {}
+
+        component.observe(lifecycleOwner, callback)
+
+        verify(idealDelegate).observe(lifecycleOwner, component.viewModelScope, callback)
+        verify(genericActionDelegate).observe(eq(lifecycleOwner), eq(component.viewModelScope), any())
+    }
+
+    @Test
+    fun `when removeObserver is called then removeObserver in delegates is called`() {
+        component.removeObserver()
+
+        verify(idealDelegate).removeObserver()
+        verify(genericActionDelegate).removeObserver()
+    }
+
+    @Test
+    fun `when component is initialized then view flow should bu null`() = runTest {
+        component.viewFlow.test {
+            assertNull(awaitItem())
+            expectNoEvents()
+        }
+    }
+
+    @Test
+    fun `when action delegate view flow emits a value then component view flow should match that value`() = runTest {
+        val actionDelegateViewFlow = MutableStateFlow(TestComponentViewType.VIEW_TYPE_1)
+        whenever(genericActionDelegate.viewFlow) doReturn actionDelegateViewFlow
+        component = IdealComponent(
+            idealDelegate,
+            genericActionDelegate,
+            actionHandlingComponent,
+            componentEventHandler,
+        )
+
+        component.viewFlow.test {
+            assertEquals(TestComponentViewType.VIEW_TYPE_1, awaitItem())
+
+            actionDelegateViewFlow.emit(TestComponentViewType.VIEW_TYPE_2)
+            assertEquals(TestComponentViewType.VIEW_TYPE_2, awaitItem())
+
+            expectNoEvents()
+        }
+    }
+}